+++ bigdbm/deliver/__init__.py
@@ -1,2 +1,2 @@
 """Convert results into CSVs and other formats."""
-from bigdbm.format.csv import CSVStringFormatter
+from bigdbm.deliver.csv import CSVStringFormatter
